Update lab dependency
Bump the lab dependency to work with the newest version.  Use a rather
permissive ‘>=0.4, <1.0’ since in all likelihood we’re going to be
compatible with all future ‘0.*’ versions and this saves us a bump each
time.  I guess we’ll see if that turns out to be a mistake.

Note that since version 0.7, lab improved accuracy of its caluclations
thus this affected some of the tests.