Rename `de2000` module to `cie00`
The `cie00` is shorter but what’s more important better indicates the
origin of the metric, i.e. that it was defined by the CIE in 2000.
Previous name only referred to the year.  While it’s unlikely that it
would lead to any confusion, it’s going to help with naming when
formulæ defined by other groups will be added.

For backwards compatibility with 0.2.2 version a hidden re-export under
the old name is still available.