bench: generate L*a*b* colours directly rather than going through sRGB
Instead of generating random 24-bit sRGB colours and than converting
them to L*a*b*, generate L*a*b* directly.  The range of L*, a* and b*
are a tad wider than the gamut of sRGB though that’s just a minor
difference.  To be honest, this change is mostly made to allow for
easier removal of the lab dependency in benchmarks and tests.