Add support for `rgb` crate
`rgb` crate provides types for holding RGB colours.  Make it possible
to pass those types (specifically `RGB<u8>` or `RGBA<u8>`) to to `diff`
methods so that we play nicely with any other users of the crate.  This
is of course done by implementing `ToLab` trait for them.

With this, deprecate `diff_rgb` in favour of calling `diff` with
`rgb::RGB<u8>` arguments.