Remove support for RGBA
It’s not obvious that alpha channel should be ignored when comparing
colours.  Perhaps a difference between fully opaque and fully
transparent colour should always be zero?  The behaviour is application-
-specific.  Get rid of RGBA support and let users choose whatever they
want.  If they want alpha to be ignored, it’s as easy as using `rgb`
method on an RGBA object.